introduce Glencoe's interfaces for event and node providers
